requirements.txt
asn1crypto==1.5.1
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
nanoid==2.0.0
outcome==1.1.0
pg8000==1.24.1
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.19.2
scramp==1.4.1
selenium==4.1.3
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.32
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.1.0
